Added abstract class CRCTable extending abstract class CRC. CRCTable generates a 256-byte CRC lookup table (GenerateTable()) and generates a CRC using the table lookup (GenerateCRC()).
In order to use the table lookup method, the order must be a factor of 8. If an attempt is made to instantiate a non-abstract subclass of CRCTable with an order not cleanly divisible by 8 and ArgumentOutOfRangeException is thrown.

Despite using a generic _message type, the elements in the message must cleanly convert to a byte. If not, then an exception will be raised.